Import Lato font for use in charts

Lato is Copyright (c) 2010-2014, Łukasz Dziedzic, with Reserved Font Name Lato (License). You can find it on Google Fonts.

font_lato == "Lato" ("Regular")

font_lato_bold == "Lato Bold"

font_lato_light == "Lato Light"

font_lato_semibold == "Lato Semibold"

import_lato()

font_lato

font_lato_bold

font_lato_light

font_lato_semibold

Format

length 1 character vector

An object of class character of length 1.

An object of class character of length 1.

An object of class character of length 1.

Details

There is an option hrbrthemes.loadfonts which -- if set to TRUE -- will call extrafont::loadfonts() to register non-core fonts with R PDF & PostScript devices. If you are running under Windows, the package calls the same function to register non-core fonts with the Windows graphics device.

Note

This will take care of ensuring PDF/PostScript usage. The location of the font directory is displayed after the base import is complete. It is highly recommended that you install them on your system the same way you would any other font you wish to use in other programs.