Import Source Sans font for use in charts

Source Sans Pro is Copyright 2010, 2012 Adobe Systems Incorporated (http://www.adobe.com/), with Reserved Font Name ‘Source’. (License). You can find it on Google Fonts.

font_ssp == "Source Sans 3" etc.

import_source_sans()

font_ssp

font_ssp_semibold

font_ssp_light

Format

length 1 character vector

An object of class character of length 1.

An object of class character of length 1.

Details

The font files shipped with this package were retrieved from https://github.com/adobe-fonts/source-sans-pro/releases/download/3.028R/source-sans-3v028R.zip.

Unfortunately there's some confusion with the name of the base font family, which is "Source Sans Pro" on my brew installed fonts on macOS, but it's "Source Sans 3" for the GitHub-installed version on a Linux machine. I have not yet figured out how to handle this.

Note

Fonts are hard and platform-compatible font usage is hard to debug.