A precise & pristine ggplot2 theme with opinionated defaults and an emphasis on typography

This is a verbatim copy of [hrbrthemes::theme_ipsum_ps] with "Source Sans Pro" dropped in.

theme_ipsum_ss(
  base_family = font_ssp,
  base_size = 10.5,
  plot_title_family = font_ssp,
  plot_title_size = 19,
  plot_title_face = "bold",
  plot_title_margin = 10,
  subtitle_family = font_ssp_light,
  subtitle_size = 13,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = font_ssp_light,
  caption_size = 10,
  caption_face = "plain",
  caption_margin = 10,
  axis_text_size = base_size,
  axis_title_family = base_family,
  axis_title_size = 11,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = margin(20, 20, 20, 20),
  panel_spacing = grid::unit(2, "lines"),
  grid_col = "#cccccc",
  grid = TRUE,
  axis_col = "#cccccc",
  axis = FALSE,
  ticks = FALSE,
  legend_position = "top",
  plot_title_position = "plot"
)

Arguments

base_family, base_size
base font family and size

plot_title_family, plot_title_face, plot_title_size, plot_title_margin
plot tilte family, face, size and margin

subtitle_family, subtitle_face, subtitle_size
plot subtitle family, face and size

subtitle_margin
plot subtitle margin bottom (single numeric value)

strip_text_family, strip_text_face, strip_text_size
facet label font family, face and size

caption_family, caption_face, caption_size, caption_margin
plot caption family, face, size and margin

axis_text_size
font size of axis text

axis_title_family, axis_title_face, axis_title_size
axis title font family, face and size

axis_title_just
axis title font justificationk one of [blmcrt]

plot_margin
plot margin (specify with ggplot2::margin)

panel_spacing
panel spacing (use unit())

grid_col
grid color

grid
panel grid (TRUE, FALSE, or a combination of X, x, Y, y)

axis_col
axis color

axis
add x or y axes? TRUE, FALSE, "xy"

ticks
ticks if TRUE add ticks

legend_position
Same as legend.position in ggplot2::theme(), defaulting to "top".

plot_title_position
Same as plot.title.position in ggplot2::theme(), defaulting to "plot".

Details

You should import_source_sans() first and also install the fonts on your system before trying to use this theme.

There is an option hrbrthemes.loadfonts which -- if set to TRUE -- will call extrafont::loadfonts() to register non-core fonts with R PDF & PostScript devices. If you are running under Windows, the package calls the same function to register non-core fonts with the Windows graphics device.

Examples

if (FALSE) {
library(ggplot2)
library(dplyr)

# seminal scatterplot
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(
    x = "Fuel efficiency (mpg)", y = "Weight (tons)",
    title = "Seminal ggplot2 scatterplot example",
    subtitle = "A plot that is only useful for demonstration purposes and the letter g",
    caption = "Brought to you by the letter 'g'"
  ) +
  theme_ipsum_ss()

# seminal bar chart

update_geom_font_defaults(family = font_ssp)

count(mpg, class) %>%
  ggplot(aes(class, n)) +
  geom_col() +
  geom_text(aes(label = n), nudge_y = 3) +
  labs(
    x = "Fuel efficiency (mpg)", y = "Weight (tons)",
    title = "Seminal ggplot2 bar chart example",
    subtitle = "A plot that is only useful for demonstration purposes",
    caption = "Brought to you by the letter 'g'"
  ) +
  theme_ipsum_ss(grid = "Y") +
  theme(axis.text.y = element_blank())
}